Add additional comments for documentation
Move FillData to its own include file so the data file include can be
moved to the top and act like a normal include file.
