+++ src/QMCDrivers/tests/convert_hdf_to_cpp.py
@@ -63,7 +63,7 @@ out_func = """
 // Generated with convert_hdf_to_cpp.py
 // clang-format off
 
-void get_diamond_fill_data(FillData &fd)
+void get_diamond_fill_data(qmcplusplus::FillData &fd)
 {
 """
